Блогикум(1)

+++ blogicum/blog/views.py
@@ -1,16 +1,14 @@
-from django.shortcuts import render, redirect, get_object_or_404
-from django.utils import timezone
 from django.contrib.auth import get_user_model, login
 from django.contrib.auth.decorators import login_required
-from django.views.generic import UpdateView, DeleteView, CreateView
-from django.contrib.auth.mixins import LoginRequiredMixin
-from django.contrib.auth.forms import UserCreationForm
-from django.urls import reverse_lazy
+from django.core.exceptions import PermissionDenied
 from django.core.paginator import Paginator
 from django.db.models import Count
+from django.http import Http404
+from django.shortcuts import get_object_or_404, redirect, render
+from django.utils import timezone
+from django.views.generic import CreateView
 
 from .models import Category, Post, Comments
-from .constants import POSTS_LIMIT
 from .forms import CreatePost, ProfileForm, CommentForm, CustomUserCreationForm
 
 User = get_user_model()
@@ -40,7 +38,8 @@ def category_posts(request, category_slug):
     )
     post_list = get_select_related(
         filter_posts_by_publication(
-            category.posts.all().annotate(comment_count=Count('comments')).order_by('-pub_date')
+            category.posts.all().annotate(
+                comment_count=Count('comments')).order_by('-pub_date')
         )
     )
     paginator = Paginator(post_list, 10)
@@ -57,9 +56,9 @@ def category_posts(request, category_slug):
 
 
 def post_detail(request, pk):
-    post = get_object_or_404(
-        filter_posts_by_publication(Post.objects.all()), pk=pk
-    )
+    post = Post.objects.filter(pk=pk).first()
+    if not post or (not post.is_published and request.user != post.author):
+        raise Http404
     comments = Comments.objects.filter(post=post)
     form = CommentForm()
     return render(
@@ -72,7 +71,8 @@ def post_detail(request, pk):
 def index(request):
     post_list = get_select_related(
         filter_posts_by_publication(
-            Post.objects.annotate(comment_count=Count('comments')).order_by('-pub_date')
+            Post.objects.annotate(
+                comment_count=Count('comments')).order_by('-pub_date')
         )
     )
     paginator = Paginator(post_list, 10)
@@ -85,14 +85,15 @@ def index(request):
     )
 
 
-@login_required
 def profile_view(request, username):
     profile = get_object_or_404(User, username=username)
-    post_list = Post.objects.filter(author=profile).annotate(comment_count=Count('comments')).order_by('-pub_date')
+    post_list = Post.objects.filter(author=profile).annotate(
+        comment_count=Count('comments')).order_by('-pub_date')
     paginator = Paginator(post_list, 10)
     page_number = request.GET.get('page')
     page_obj = paginator.get_page(page_number)
-    return render(request, 'blog/profile.html', {'profile': profile, 'page_obj': page_obj})
+    return render(request, 'blog/profile.html',
+                  {'profile': profile, 'page_obj': page_obj})
 
 
 @login_required
@@ -124,29 +125,38 @@ def edit_profile(request):
     return render(request, 'blog/user.html', {'form': form})
 
 
-class PostUpdateView(LoginRequiredMixin, UpdateView):
-    model = Post
-    form_class = CreatePost
-    template_name = 'blog/create.html'
-
-    def get_success_url(self):
-        return reverse_lazy('blog:post_detail', kwargs={'pk': self.object.pk})
+@login_required
+def edit_post(request, pk):
+    post = get_object_or_404(Post, pk=pk)
+    if post.author != request.user:
+        return redirect('blog:post_detail', pk=post.pk)
+    if request.method == 'POST':
+        form = CreatePost(request.POST, instance=post)
+        if form.is_valid():
+            form.save()
+            return redirect('blog:post_detail', pk=post.pk)
+    else:
+        form = CreatePost(instance=post)
+    return render(request, 'blog/create.html', {'form': form})
 
 
 @login_required
 def delete_post(request, pk):
     post = get_object_or_404(Post, pk=pk)
-    form = CreatePost(instance=post)
+    if request.user != post.author and not request.user.is_staff:
+        raise PermissionDenied
     if request.method == 'POST':
         post.delete()
         return redirect('blog:profile', username=request.user)
-    return render(request, 'blog/create.html', {'form': form})
+    return render(request, 'blog/create.html', {'post': post})
 
 
 @login_required
 def edit_comment(request, post_id, comment_id):
     post = get_object_or_404(Post, pk=post_id)
     comment = get_object_or_404(Comments, pk=comment_id, post=post)
+    if comment.author != request.user:
+        raise PermissionDenied
     if request.method == 'POST':
         form = CommentForm(request.POST, instance=comment)
         if form.is_valid():
@@ -157,7 +167,8 @@ def edit_comment(request, post_id, comment_id):
             return redirect('blog:post_detail', pk=post_id)
     else:
         form = CommentForm(instance=comment)
-    return render(request, 'blog/comment.html', {'post': post, 'form': form, 'comment': comment})
+    return render(request, 'blog/comment.html',
+                  {'post': post, 'form': form, 'comment': comment})
 
 
 @login_required
@@ -173,17 +184,23 @@ def add_comment(request, pk):
             return redirect('blog:post_detail', pk=pk)
     else:
         form = CommentForm()
-    return render(request, 'blog/comment.html', {'post': post, 'form': form, 'comments': post.comments.all()})     
+    return render(
+        request, 'blog/comment.html',
+        {'post': post, 'form': form, 'comments': post.comments.all()}
+    )
 
 
 @login_required
 def delete_comment(request, post_id, comment_id):
     post = get_object_or_404(Post, pk=post_id)
     comment = get_object_or_404(Comments, pk=comment_id, post=post)
+    if request.user != comment.author and not request.user.is_staff:
+        raise PermissionDenied
     if request.method == 'POST':
         comment.delete()
         return redirect('blog:post_detail', pk=post_id)
-    return render(request, 'blog/comment.html', {'post': post, 'comment': comment})
+    return render(request, 'blog/comment.html',
+                  {'post': post, 'comment': comment})
 
 
 class UserCreateView(CreateView):
